# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Plant nursery store - OpenAPI 3.0
  description: |-
    This is a sample Plant Nursery Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). 

  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://xeniya-dob-plants-nursery-store.onrender.com/api/v1
...
paths:
...
  /items:
    post:
      tags:
      - Items
      summary: Create Single Item
      operationId: CreateSingleItem
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateSingleItemRequest'
              - example:
                  name: Pine
                  price: 3
                  description: Lorem
            example:
              name: Pine
              price: 3
              description: Lorem
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    get:
      tags:
      - Items
      summary: Get All Items
      operationId: GetAllItems
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    CreateSingleItemRequest:
      title: CreateSingleItemRequest
      required:
      - name
      - price
      - description
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        description:
          type: string          
      example:
        name: Jasmine
        price: 10
        description: Lorem
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []